tools/piano_vision_training/requirements.txt
ultralytics==8.3.0
onnx==1.17.0
onnxruntime==1.20.1
numpy==2.1.3
opencv-python==4.10.0.84
PyYAML==6.0.2
matplotlib==3.9.2
